* removed the global uiFrontbuf from interface.c (and python, it was used   there) * replaced with nice local (uiBlock) storage of what goes on in drawing.   it now only calls a glDrawBuffer() and glFinish() when it is actually   needed
Result: interface drawing in general is speedy again, especially for gfx
cards that dont allow frontbuffer drawing, and copy stuff to the frontbuf
with a glFinish() call.

Needs to be tested on all platforms... report to me when you see problems
like menus not drawing correctly, tooltips not drawing or not disappearing,
etc.