Bug fix #1477
Important for everyone to know; buttons with event return code 0 pass on
the button event to other buttons. This was in yafray panels so, causing
buttons unwanted to be pressed when dragging sliders.

The official 'empty' event is B_DIFF, to denote that back and frontbuffers
are different, since the button was updated in front only with a press.
Alternative is B_NOP, which doesnt do anything.