Changed type for new textimeout and texcollectrate into "int".
Please note that the "long" should be avoided in dna structs... it is a
reserved type for use in our code for pointer storage. It is 32 or 64 bits,
depending CPU architecture (pointer size).

The "long" type currently is in the makesdna.c code, and interpreted as 32
bits only, which is actually wrong yes... I should remove it or code it
correct.

BTW; this gave a warning in space.c, another case where MSVC forgives you?
:)