(Orange branch) Phew... a whole day revision of library linking system... there was a bug when objects got linked, then removed, and then linked via a group... this was a weak part in the library system, when data is used both "direct" (like local object is linked to scene) and "indirect" (when group from a library has same object).
The crucial fixes are easier to find by reading the diff... this system
is just braincrunching complex! Just as note to self then;
- ID_ID oldmap-lookups are now protected
- removed adding oldmap lookups for already read data

For Matt & Basse: you can do a very good cleanup of old files by putting
the "rt" button on 127, and then save file. This will skip writing any
reference to library data in the file, with exception of group links.
Reload a file might give some link errors, but a 2nd save and load then
should result in a clean file.