Bugfix: (own todo)
When you rename a generated image, like "Render Result" or "Composite"
the float buffer in this image had to be copied, otherwise a next render
will make the renamed Image buffer invalid.

Note that renaming will not mean the images get packed or saved! So on
reloading .blend they're gone.