Fix for bug #5910: Saving in sculpt mode with partial visibility would give "not enough diskspace" errors. That wasn't the error at all though, but blender gives this diskspace error message on any error (this is for another commit to fix though).
The cause of the error was that with partial visibility it saved the
MEdge and MFace arrays as if they had the full length, which is not
the case, so it was reading past the end of the array. I added an
exception for these.

Also removed the oldstyle tface saving, this was only there to provide
temporary upward compatibility until release.