=== Transform Snap ===
[ #6131 ] Vertex snapping broken in shaded face + edge select mode

The problem was trying to use the OGL optimisation when vertex selection was turned off. Added a check for that in all the find_nearest functions (vert, edge, face) so nobody falls in that trap again.

The offshot is that snapping in shaded view without vertex select on can snap to occluded vertice.