== Interface == * Tablet pressure sensitivity for number field dragging
Many of the number fields in Blender are very sensitive. With this addition, 
softer tablet pressure causes the number field dragging to be more precise, 
as if you're holding shift, but variable depending on how hard or soft 
you're pressing. You can push hard for large adjustments or just lightly
stroke it for fine tuning.

P.S. There was a bug in the tracker regarding tablet support in sculpt mode that
never really got resolved. I don't know if it still exists, but number fields 
are a pretty major part of Blender, so just in case this commit causes problems
for you, you can disable it (temporarily, this won't be left in) by changing
rt: to any value other than 0. If anyone has problems, please report them, and
we can get it sorted out properly.