This commit does a couple of things to the Makefiles.
The first is it adds libIlmThread.a to the OpenEXR libs.

The second thing it does, is make it possible to define what TARGETS
you want to build inside of your user-def.mk file.  This simplifies
source/Makefile quite a bit.  I made each platform have the same
defaults (build dynamic blender, gameengine and gameplayer)  I think
most platforms should be building this combo anyway and if you need to
change it its trivial so no need to have different defaults for each platform.
The new defines are as follows with their default settings in:
         source/nan_definitions.mk

    export WITH_BF_DYNAMICOPENGL ?= true
    export WITH_BF_STATICOPENGL ?= false
    export WITH_BF_GAMEENGINE ?= true
    export WITH_BF_GAMEPLAYER ?= true
    export WITH_BF_WEBPLUGIN ?= false

Let me know if you have any problems with it.

Kent