Patch #6140
Transform Properties panel in 3d window.
New button "Parbone" showing when parent relation is to a specific bone.
This information was nowhere reveiled in the UI until now.

Code remark:
The code isn't getting more beautiful now with all if() options added.
The idea to move the rotation buttons down when a parent bone was set isn't
very good, buttons should stay as much as possible on the same location.
However, I dont have time for a solution either...