only draw the xor line when the mouse moves (border moving). This is the only thing that made software GL unuseable, and would even hang blender when using some debugging tools. The lag was noticibel on some cheap graphics cards also.
There is a problem with blenders events that means the mouse movement events are sent while the mouse button release event isnt for some time. This could make resizing a border take a fairly long time... 4-10sec.
even though this is a workaround, only redrawing when the mouse is moved is logical.