This fixes make release so that it does not package .svn dirs with the build. This is bug [#7387] Release files pack the .svn metadata directories
I took a breif look at the scons files and it looks like it has a fix
for them already.  It would be great if someone that uses scons could
just do a quick double check and let me know if there are problems with it
and or just fix whatever slips through.

Kent