Nodes to be created at the location in which the mouse cursor was initially This commit changes the way new nodes are created. Previously nodes were created at the last location of the mouse cursor before creating a new node. After this commit the nodes are created at the initial location in which the cursor was before the menu used to add the node was opened.
This makes it possible for the user to predict where the new node appears without
having to move it to desired location like before. There is no option for this as
this behaviour is clearly superior to previous.

Codewise it uses static variables. This is something to fix during 2.50 rewrite.

Thanks to paprmh for contribution!