Patch #21911: Uppercase Windows Drive Letters in File Open Window Submitted by: Harley Acheson
Harley writes:
The normal convention in Windows is to show drive letters in uppercase, as in "C:\", not "c:\" as
seen in the File window.

Although Windows file paths are generally case insensitive, drive letters have been shown as uppercase since the early
DOS days.  Seeing them in lowercase in the Blender File window looks a bit strange.