== rna cleanup script ==
- new folder rna_cleanup/
- moved na_api_cleanup.txt in rna_cleanup/
- rna_cleaner.py is a script to help cleaning rna names, pasting the help below for those interested

Basically after you run this on a original file, the script produces 2 file .txt and .py that you can edit.
You can skip to edit things liek "changed" or "same" or check if the "to" field is correct or not.
When you re-run this script it will check these thigns and will produce a consistent output file again.

Also, you can sort lines at will.

$ ./rna_cleaner.py -h

HELP:
Run this script to re-format the edits you make in the input file.
Do quick modification to important fields like 'to' and don't care about fields like 'changed' or 'description' and save.
The script outputs 3 files:
   1) *_clean.txt: is formatted same as the .txt input, can be edited by user.
   2) *_clean.py: is formatted same as the .py input, can be edited by user.
   3) rna_api.py is not formatted for readability and go under complete check. Can be used for rna cleanup.


USAGE:
./rna_cleaner.py input-file (.txt|.py) order-priority (note|changed|class|from|to|kw).
./rna_cleaner.py -h for help