Bugfix #25086
The texture node is procedural; like RGB node it has no real buffers,
but allows per-pixel reading. The compositor uses nodes that directly
access buffers too, which conflicts with it... needs more design here.

Restored old functionality that just passes on preview size buffers
for nodes to prevent crashes. Giving it a render-size buffer is not
nice; the resolution-independence of texture nodes is interesting to
keep.

Solution could be:
- visually tag input/output sockets for this case (sockets with buffers, 
  vs sockets with values), so users know what to expect.