Bugfix #25551
Transform crash on autokey during animplay.
Transform event handling requires Context, for inserting keys. 

Marked it with XXX warning; we have to be more careful with which
API calls require context; like inserting keys shouldn't need it?

Originally Context was meant for operators; to define user level
or scripted context, not for underlying APIs to work.