Color management support for sequencer
This adds color management controls to the sequencer space,
which are applying on the displaying image.

Currently used the same UI approach as was used for other
areas, but this would probably be cleaned up in the nearest
future.

This required a bit of hackish stuff to be done, Namely color
management for the sequencer requires one extra step in the
conversion chain to convert sRGB float buffer to linear space.
This was made by setting special flag in ImBuf. Later this
could be done using input color space for the ImBuf probably.