Tomato Cycles: report currently rendering sample when using GPU
Currently it makes more sense to use single tile for GPU rendering
and in this case tile-based progress report doesn't work well.

Since threading happens within single tile, it's possible to
detect currently computing sample and report it to the interface,

This also allows to display smoother progress when using CPU
with small amount of tiles.