Fix: multisample viewport drawing didn't work well with selection or particle brushes, due to issues with color coded drawing or slow/buggy reading from such a buffer on some systems.
In case multisample is enabled now, it uses an offscreen buffer for such drawing,
which is not multisampled and so should not cause issues. This does mean there is
some extra GPU memory usage when multisample is enabled, and we could optimize
triple buffer to work together here somehow to share buffers, but it's better than
having selection not working.