Bevel - better adj mesh pattern, and add profile control
Summary:
The 'adj mesh' pattern used an 'average position' hack to set
the location of internal vertices; this could lead to overlaps
and uneven looking corner bevels. This change switches to using
cubic subdivision with border modifications in order to make the
pattern.

Also added some profile control: rather than assuming the profile
for multi-segment bevels should follow a quarter circle, it now
uses a superellipse for the profile. The superellipse has a parameter
that is an exponent (power): when it is 2 you get circles, when it
is 1 you get a straight line, and other numbers give more convex
and more concave profiles.

Also added a special case: if the vertex is close to looking like
a cube corner, the adj mesh pattern is snapped to a superellipsoid
at the corner, giving nice sphere octant curves when the power is 2
and the faces all make 90 degree angles with each other.

Reviewers: campbellbarton, carter2422

Differential Revision: https://developer.blender.org/D164